+++ doc/release-notes.md
@@ -7,3 +7,4 @@ Upstream Bitcoin-ABC updates
 - Merge #12630: Provide useful error message if datadir is not writable
  - Using addresses in createmultisig is now deprectated. Use -deprecatedrpc=createmultisig to get the old behavior.
  - The `createrawtransaction` RPC will now accept an array or dictionary (kept for compatibility) for the `outputs` parameter. This means the order of transaction outputs can be specified by the client.
+ - The new RPC `testmempoolaccept` can be used to test acceptance of a transaction to the mempool without adding it.